// client/src/api/generated/models/VehicleSummaryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { VehicleResponseDto } from './VehicleResponseDto';
export type VehicleSummaryDto = {
    vehicle: VehicleResponseDto;
    tripCount: number;
    totalDistanceKm: number;
    totalFuelConsumed: number;
    averageDurationMinutes: number;
};